+++ src/TicTacToe.java
@@ -14,27 +14,27 @@ public class TicTacToe implements ActionListener {
     JButton[] jButtons = new JButton[9];
     boolean player1_turn;
 
-    // Frame size (TODO: could be a constant size for both the width and height which is a choice to make later on...)
+    // Frame size
     int width = 800;
     int height = 800;
 
-    // Background colour using RGB colour code (TODO: Make colour scheme match the scriptjumper colour scheme or something close...)
+    // Background colour using RGB colour code (#323232 - Dark Grey Almost Black)
     int pane_bg_red = 50;
     int pane_bg_green = 50;
     int pane_bg_blue = 50;
 
-    // Text field background colour using RGB colour code (TODO: Make colour scheme match the scriptjumper colour scheme or something close...)
+    // Text field background colour using RGB colour code (#191919 - Black)
     int textField_bg_red = 25;
     int textField_bg_green = 25;
     int textField_bg_blue = 25;
 
-    // Text field foreground colour using RGB colour code (TODO: Make colour scheme match the scriptjumper colour scheme or something close...)
+    // Text field foreground colour using RGB colour code (#1aff00 - Green/Light Green)
     int textField_fg_red = 25;
     int textField_fg_green = 255;
     int textField_fg_blue = 0;
 
     // Font styling
-    String fontName = "TimesRoman"; // TODO: will change to a personal favorite
+    String fontName = "Sathu";
     int fontStyle = Font.BOLD; // 1
     int fontSize = 75;
 
@@ -57,8 +57,6 @@ public class TicTacToe implements ActionListener {
     int grid_bg_blue = 150;
 
     // Grid Buttons
-    String buttonFontName = "TimesRoman";
-    int buttonFontStyle = Font.BOLD;
     int buttonFontSize = 120;
 
     // Players
@@ -89,7 +87,7 @@ public class TicTacToe implements ActionListener {
         for(int i = 0; i < 9; i++) {
             jButtons[i] = new JButton();
             jPanel_button.add(jButtons[i]);
-            jButtons[i].setFont(new Font(buttonFontName, buttonFontStyle, buttonFontSize));
+            jButtons[i].setFont(new Font(fontName, fontStyle, buttonFontSize));
             jButtons[i].setFocusable(false);
             jButtons[i].addActionListener(this);
         }
@@ -121,7 +119,7 @@ public class TicTacToe implements ActionListener {
             jButtons[idx].setText(playerText);
             player1_turn = isPlayerTurn;
             jLabel_textField.setText(nextPlayerText + " turn");
-            checkWinningConditions();
+            checkWinningConditions(playerText);
         }
     }
 
@@ -145,133 +143,98 @@ public class TicTacToe implements ActionListener {
 
     // Check if there is any winning conditions
     // Or if any player(X or O) has won
-    public void checkWinningConditions() {
-        // check X win conditions
+    public void checkWinningConditions(String playerText) {
         if(
-                (Objects.equals(jButtons[0].getText(), x)) &&
-                        (Objects.equals(jButtons[1].getText(), x)) &&
-                        (Objects.equals(jButtons[2].getText(), x))
+                (Objects.equals(jButtons[0].getText(), playerText)) &&
+                        (Objects.equals(jButtons[1].getText(), playerText)) &&
+                        (Objects.equals(jButtons[2].getText(), playerText))
         ) {
-            winningPlayer(0, 1, 2, x);
+            endGameMessage(0, 1, 2, playerText);
         }
         if(
-                (Objects.equals(jButtons[3].getText(), x)) &&
-                        (Objects.equals(jButtons[4].getText(), x)) &&
-                        (Objects.equals(jButtons[5].getText(), x))
+                (Objects.equals(jButtons[3].getText(), playerText)) &&
+                        (Objects.equals(jButtons[4].getText(), playerText)) &&
+                        (Objects.equals(jButtons[5].getText(), playerText))
         ) {
-            winningPlayer(3, 4, 5, x);
+            endGameMessage(3, 4, 5, playerText);
         }
         if(
-                (Objects.equals(jButtons[6].getText(), x)) &&
-                        (Objects.equals(jButtons[7].getText(), x)) &&
-                        (Objects.equals(jButtons[8].getText(), x))
+                (Objects.equals(jButtons[6].getText(), playerText)) &&
+                        (Objects.equals(jButtons[7].getText(), playerText)) &&
+                        (Objects.equals(jButtons[8].getText(), playerText))
         ) {
-            winningPlayer(6, 7, 8, x);
+            endGameMessage(6, 7, 8, playerText);
         }
         if(
-                (Objects.equals(jButtons[0].getText(), x)) &&
-                        (Objects.equals(jButtons[3].getText(), x)) &&
-                        (Objects.equals(jButtons[6].getText(), x))
+                (Objects.equals(jButtons[0].getText(), playerText)) &&
+                        (Objects.equals(jButtons[3].getText(), playerText)) &&
+                        (Objects.equals(jButtons[6].getText(), playerText))
         ) {
-            winningPlayer(0, 3, 6, x);
+            endGameMessage(0, 3, 6, playerText);
         }
         if(
-                (Objects.equals(jButtons[1].getText(), x)) &&
-                        (Objects.equals(jButtons[4].getText(), x)) &&
-                        (Objects.equals(jButtons[7].getText(), x))
+                (Objects.equals(jButtons[1].getText(), playerText)) &&
+                        (Objects.equals(jButtons[4].getText(), playerText)) &&
+                        (Objects.equals(jButtons[7].getText(), playerText))
         ) {
-            winningPlayer(1, 4, 7, x);
+            endGameMessage(1, 4, 7, playerText);
         }
         if(
-                (Objects.equals(jButtons[2].getText(), x)) &&
-                        (Objects.equals(jButtons[5].getText(), x)) &&
-                        (Objects.equals(jButtons[8].getText(), x))
+                (Objects.equals(jButtons[2].getText(), playerText)) &&
+                        (Objects.equals(jButtons[5].getText(), playerText)) &&
+                        (Objects.equals(jButtons[8].getText(), playerText))
         ) {
-            winningPlayer(2, 5, 8, x);
+            endGameMessage(2, 5, 8, playerText);
         }
         if(
-                (Objects.equals(jButtons[0].getText(), x)) &&
-                        (Objects.equals(jButtons[4].getText(), x)) &&
-                        (Objects.equals(jButtons[8].getText(), x))
+                (Objects.equals(jButtons[0].getText(), playerText)) &&
+                        (Objects.equals(jButtons[4].getText(), playerText)) &&
+                        (Objects.equals(jButtons[8].getText(), playerText))
         ) {
-            winningPlayer(0, 4, 8, x);
+            endGameMessage(0, 4, 8, playerText);
         }
         if(
-                (Objects.equals(jButtons[2].getText(), x)) &&
-                        (Objects.equals(jButtons[4].getText(), x)) &&
-                        (Objects.equals(jButtons[6].getText(), x))
+                (Objects.equals(jButtons[2].getText(), playerText)) &&
+                        (Objects.equals(jButtons[4].getText(), playerText)) &&
+                        (Objects.equals(jButtons[6].getText(), playerText))
         ) {
-            winningPlayer(2, 4, 6, x);
+            endGameMessage(2, 4, 6, playerText);
         }
 
-        // Check O win conditions
-        if(
-                (Objects.equals(jButtons[0].getText(), o)) &&
-                        (Objects.equals(jButtons[1].getText(), o)) &&
-                        (Objects.equals(jButtons[2].getText(), o))
-        ) {
-            winningPlayer(0, 1, 2, o);
-        }
-        if(
-                (Objects.equals(jButtons[3].getText(), o)) &&
-                        (Objects.equals(jButtons[4].getText(), o)) &&
-                        (Objects.equals(jButtons[5].getText(), o))
-        ) {
-            winningPlayer(3, 4, 5, o);
-        }
-        if(
-                (Objects.equals(jButtons[6].getText(), o)) &&
-                        (Objects.equals(jButtons[7].getText(), o)) &&
-                        (Objects.equals(jButtons[8].getText(), o))
-        ) {
-            winningPlayer(6, 7, 8, o);
-        }
-        if(
-                (Objects.equals(jButtons[0].getText(), o)) &&
-                        (Objects.equals(jButtons[3].getText(), o)) &&
-                        (Objects.equals(jButtons[6].getText(), o))
-        ) {
-            winningPlayer(0, 3, 6, o);
-        }
-        if(
-                (Objects.equals(jButtons[1].getText(), o)) &&
-                        (Objects.equals(jButtons[4].getText(), o)) &&
-                        (Objects.equals(jButtons[7].getText(), o))
-        ) {
-            winningPlayer(1, 4, 7, o);
-        }
-        if(
-                (Objects.equals(jButtons[2].getText(), o)) &&
-                        (Objects.equals(jButtons[5].getText(), o)) &&
-                        (Objects.equals(jButtons[8].getText(), o))
-        ) {
-            winningPlayer(2, 5, 8, o);
-        }
-        if(
-                (Objects.equals(jButtons[0].getText(), o)) &&
-                        (Objects.equals(jButtons[4].getText(), o)) &&
-                        (Objects.equals(jButtons[8].getText(), o))
-        ) {
-            winningPlayer(0, 4, 8, o);
-        }
-        if(
-                (Objects.equals(jButtons[2].getText(), o)) &&
-                        (Objects.equals(jButtons[4].getText(), o)) &&
-                        (Objects.equals(jButtons[6].getText(), o))
-        ) {
-            winningPlayer(2, 4, 6, o);
+        if (!Objects.equals(jButtons[0].getText(), "") &&
+                !Objects.equals(jButtons[1].getText(), "") &&
+                !Objects.equals(jButtons[2].getText(), "") &&
+                !Objects.equals(jButtons[3].getText(), "") &&
+                !Objects.equals(jButtons[4].getText(), "") &&
+                !Objects.equals(jButtons[5].getText(), "") &&
+                !Objects.equals(jButtons[6].getText(), "") &&
+                !Objects.equals(jButtons[7].getText(), "") &&
+                !Objects.equals(jButtons[8].getText(), "")) {
+            endGameMessage(0, 0, 0, "");
         }
     }
 
-    public void winningPlayer(int a, int b, int c, String player) {
-        jButtons[a].setBackground(Color.GREEN);
-        jButtons[b].setBackground(Color.GREEN);
-        jButtons[c].setBackground(Color.GREEN);
+    public void endGameMessage(int a, int b, int c, String player) {
+        String endGameMessage;
+
+        if (!Objects.equals(player, "")) {
+            endGameMessage = player + " wins";
+
+            // Setting the background color of the winning combination
+            jButtons[a].setBackground(Color.BLUE);
+            jButtons[a].setOpaque(true);
+            jButtons[b].setBackground(Color.BLUE);
+            jButtons[b].setOpaque(true);
+            jButtons[c].setBackground(Color.BLUE);
+            jButtons[c].setOpaque(true);
+        } else {
+            endGameMessage = "Draw";
+        }
 
         for(int i = 0; i < 9; i++) {
             jButtons[i].setEnabled(false);
         }
 
-        jLabel_textField.setText(player + " wins");
+        jLabel_textField.setText(endGameMessage);
     }
 }